+++ homebytwo/routes/tests/test_prediction_model.py
@@ -1,7 +1,11 @@
+from django.core.exceptions import FieldError
 from django.urls import reverse
+
+import pytest
 
 from ..forms import ActivityPerformanceForm
 from ..models import ActivityType
+from ..models.activity import ActivityPerformance, PredictedModel
 from ..prediction_model import PredictionModel
 from .factories import (
     ActivityFactory,
@@ -10,6 +14,71 @@ from .factories import (
     GearFactory,
     RouteFactory,
 )
+
+
+@pytest.fixture
+def predicted_activity_types(athlete):
+    ActivityFactory.create_batch(5, athlete=athlete)
+    for activity_type in ActivityType.objects.all():
+        activity_type.train_prediction_model()
+
+
+def test_predicted_model_init(athlete):
+    activity_type = ActivityType()
+    assert isinstance(activity_type, PredictedModel)
+    assert activity_type._activity_type is activity_type
+    assert len(activity_type.regression_coefficients) == 6
+
+    activity_performance = ActivityPerformance(
+        athlete=athlete, activity_type=ActivityTypeFactory()
+    )
+    assert isinstance(activity_performance, PredictedModel)
+    assert activity_performance._activity_type is activity_performance.activity_type
+    assert len(activity_performance.regression_coefficients) == 7
+
+
+@pytest.mark.django_db
+def test_predicted_model_defaults():
+    activity_type = ActivityTypeFactory(name=ActivityType.ROLLERSKI)
+    activity_performance = ActivityPerformanceFactory()
+
+    assert len(activity_type.regression_coefficients) == 6
+    assert len(activity_performance.regression_coefficients) == 7
+
+
+def test_predicted_model_init_no_activity_type():
+    class BadModel(PredictedModel):
+        def get_training_activities(self, max_num_activities):
+            pass
+
+    with pytest.raises(FieldError):
+        BadModel()
+
+
+def test_get_training_data_activity_type(athlete):
+    limit = 5
+    activity_type = ActivityTypeFactory()
+    ActivityFactory.create_batch(10, athlete=athlete, activity_type=activity_type)
+
+    limited_activities = activity_type.get_training_activities(limit=limit)
+    assert limited_activities.count() == limit
+    assert list(activity_type.get_training_activities()) == list(
+        activity_type.activities.all()
+    )
+
+
+def test_get_training_data_activity_performance(athlete):
+    limit = 5
+    activity_performance = ActivityPerformanceFactory(athlete=athlete)
+    ActivityFactory.create_batch(
+        10, athlete=athlete, activity_type=activity_performance.activity_type
+    )
+
+    limited_activities = activity_performance.get_training_activities(limit=limit)
+    assert limited_activities.count() == limit
+    assert list(activity_performance.get_training_activities()) == list(
+        activity_performance.activity_type.activities.all()
+    )
 
 
 def test_prediction_model_with_defaults():
@@ -37,10 +106,25 @@ def test_prediction_model_with_custom_parameters():
     prediction_model = PredictionModel(
         categorical_columns=["gear", "workout_type"],
         numerical_columns=[],
+        polynomial_columns=[],
     )
 
     assert prediction_model.onehot_encoder_categories == "auto"
     assert prediction_model.numerical_columns == []
+    assert prediction_model.polynomial_columns == []
+
+
+def test_train_prediction_model(athlete):
+    activity_type = ActivityTypeFactory()
+    performance = ActivityPerformanceFactory(
+        athlete=athlete, activity_type=activity_type
+    )
+    activity = ActivityFactory(athlete=athlete, activity_type=activity_type)
+    result = performance.train_prediction_model()
+
+    assert "successfully trained" in result
+    assert performance.gear_categories == [activity.gear.strava_id]
+    assert performance.workout_type_categories == [activity.get_workout_type_display()]
 
 
 def test_train_prediction_model_data_no_data(athlete):
@@ -48,20 +132,6 @@ def test_train_prediction_model_data_no_data(athlete):
     activity_type = activity_performance.activity_type.name
     result = activity_performance.train_prediction_model()
     assert f"No training data found for activity type: {activity_type}" in result
-
-
-def test_train_prediction_model_data_success(athlete):
-    activity_performance = ActivityPerformanceFactory(athlete=athlete)
-    activity = ActivityFactory(
-        athlete=athlete, activity_type=activity_performance.activity_type
-    )
-    result = activity_performance.train_prediction_model()
-
-    assert "Model successfully trained" in result
-    assert activity_performance.gear_categories == [activity.gear.strava_id]
-    assert activity_performance.workout_type_categories == [
-        activity.get_workout_type_display()
-    ]
 
 
 def test_train_prediction_model_data_default_run(athlete):
@@ -73,7 +143,7 @@ def test_train_prediction_model_data_default_run(athlete):
     )
     result = activity_performance.train_prediction_model()
 
-    assert "Model successfully trained" in result
+    assert "successfully trained" in result
     assert activity_performance.gear_categories == [activity.gear.strava_id]
     assert activity_performance.workout_type_categories == [
         activity.get_workout_type_display()
@@ -89,7 +159,7 @@ def test_train_prediction_model_data_success_no_gear_no_workout_type(athlete):
         workout_type=None,
     )
     result = activity_performance.train_prediction_model()
-    assert "Model successfully trained" in result
+    assert "successfully trained" in result
     assert activity_performance.gear_categories == ["None"]
     assert activity_performance.workout_type_categories == ["None"]
 
@@ -161,27 +231,28 @@ def test_activity_performance_form(athlete):
     ]
 
 
-def test_activity_performance_form_no_activity_performance(athlete):
-    athlete_activity_type, other_activity_type = ActivityTypeFactory.create_batch(2)
+def test_activity_performance_form_no_performance(athlete, predicted_activity_types):
+    athlete_activity_type = ActivityType.objects.first()
+    other_activity_type = ActivityType.objects.last()
     route = RouteFactory(activity_type=other_activity_type)
+    ActivityFactory(athlete=athlete, activity_type=athlete_activity_type)
     ActivityPerformanceFactory(
         athlete=athlete,
         activity_type=athlete_activity_type,
     )
     form = ActivityPerformanceForm(route=route, athlete=athlete)
-
-    assert len(form.fields["activity_type"].choices) == len(
-        ActivityType.SUPPORTED_ACTIVITY_TYPES
+    assert (
+        len(form.fields["activity_type"].choices) == athlete.performances.all().count()
     )
     assert "gear" not in form.fields
-    assert "workout_type" not in form.fields
 
 
 def test_activity_performance_form_not_logged_in(athlete):
     form = ActivityPerformanceForm(route=RouteFactory(), athlete=None)
 
-    assert len(form.fields["activity_type"].choices) == len(
-        ActivityType.SUPPORTED_ACTIVITY_TYPES
+    assert (
+        len(form.fields["activity_type"].choices)
+        == ActivityType.objects.predicted().count()
     )
     assert "gear" not in form.fields
     assert "workout_type" not in form.fields
@@ -217,14 +288,16 @@ def test_performance_form_on_route_page(athlete, client):
         activity_performance,
         other_activity_performance,
         *_,
-    ) = ActivityPerformanceFactory.create_batch(8, athlete=athlete)
+    ) = ActivityPerformanceFactory.create_batch(3, athlete=athlete)
+    for performance in ActivityPerformance.objects.all():
+        ActivityFactory(activity_type=performance.activity_type, athlete=athlete)
     route = RouteFactory(activity_type=activity_performance.activity_type)
     url = reverse("routes:route", kwargs={"pk": route.pk})
     selected_activity_type = other_activity_performance.activity_type
     data = {"activity_type": selected_activity_type}
     response = client.post(url, data=data)
 
-    athlete_activity_types = athlete.activityperformance_set.all()
+    athlete_activity_types = athlete.performances.all()
 
     selected = '<option value="{}" selected>{}</option>'.format(
         selected_activity_type.name, selected_activity_type.get_name_display()